requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
Flask-Login>=0.5
Flask-WTF>=1.0
WTForms[email]
Werkzeug>=2.0  # For password hashing and file handling
pandas>=1.3
openpyxl>=3.0  # Needed by pandas to read .xlsx files
Pillow>=8.0    # For image handling (optional, for resizing avatars etc.)
python-dotenv>=0.19 # To load environment variables for config (optional but good practice)